+++ test/github/copilot_sdk/send_and_wait_test.clj
@@ -319,4 +319,3 @@
           (is (= :threw (first result)))
           (is (re-find #"Session has been disconnected" (second result))
               "the outcome must be a consistent disconnected error"))))))
-